Refactor similar ActivityIndicators by using initializer
